requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.8
django-celery-beat==2.6.0
django-timezone-field==7.0
dropbox==12.0.2
flower==2.0.1
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
humanize==4.10.0
idna==3.7
kombu==5.4.0
oauthlib==3.2.2
ply==3.11
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stone==3.3.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
